Added ListOfRegulars class, and updated others
ListOfRegulars, can be used to return all registered regulars, and sort them by their wealth, this is used to calculate how well a player did in the game.
Also have implemented this class in most all other classes (DanceMech, PrivateRoom, such..)
Also added a Hotel/Motel invite class, that will call different methods according to the wealth of the regular. Calculates if you are invited to a hotel or a motel, this will have an influence on your win grade.